+++ NEWS.md
@@ -3,6 +3,9 @@
 * Updated for compatibility with major `ggplot2` update; back-compatibility maintained.
 * Restored `taxize` functionality as a Suggested package.
 * Improved test coverage.
+* Moved `waffle` to Suggested package, as it is now orphaned.
+* `plot.occCiteData()` now automatically skips source and aggregator plots if `waffle` is unavailable.
+* Small bug fixes.
 
 # occCite 0.5.9